src/overlay-manager/requirements.txt
Flask==3.0.2
pika==1.3.2
PyYAML==6.0.1
PyMySQL==1.1.0
